+++ docs/changelog.md
@@ -1,6 +1,8 @@
 # 0.2.0
 - [X] Adicionado a criação do usuário
 - [X] Adicionado a configuração base do auth server
+- [X] Adicionado a customização do token para carregar as roles
+- [X] Adicionado o user details para carregar o usuário
 
 # 0.1.0
 - [X] Adicionado o arquivo de changelog